// client/src/interfaces.tsx
export interface ISetItem {
  index: String;
  ID: String;
  set: String;
  rarity: String;
  lvl?: String;
  lvl_req: String;
  ItemName: String;
  item?: String;
  found: String;
  prop1?: String;
  min1?: String;
  max1?: String;
  stats1?: String;
  prop2?: String;
  min2?: String;
  max2?: String;
  stats2?: String;
  prop3?: String;
  min3?: String;
  max3?: String;
  stats3?: String;
  prop4?: String;
  min4?: String;
  max4?: String;
  stats4?: String;
  prop5?: String;
  min5?: String;
  max5?: String;
  stats5?: String;
  prop6?: String;
  min6?: String;
  max6?: String;
  stats6?: String;
  prop7?: String;
  min7?: String;
  max7?: String;
  stats7?: String;
  prop8?: String;
  min8?: String;
  max8?: String;
  stats8?: String;
  prop9?: String;
  min9?: String;
  max9?: String;
  stats9?: String;
  prop10?: String;
  min10?: String;
  max10?: String;
  stats10?: String;
  lineString: String;
}

export interface IUniqueItem {
  index: String;
  ID: String;
  version: String;
  enabled: String;
  rarity: String;
  lvl?: String;
  lvl_req: String;
  code?: String;
  ItemName: String;
  found: String;
  prop1?: String;
  min1?: String;
  max1?: String;
  stats1?: String;
  prop2?: String;
  min2?: String;
  max2?: String;
  stats2?: String;
  prop3?: String;
  min3?: String;
  max3?: String;
  stats3?: String;
  prop4?: String;
  min4?: String;
  max4?: String;
  stats4?: String;
  prop5?: String;
  min5?: String;
  max5?: String;
  stats5?: String;
  prop6?: String;
  min6?: String;
  max6?: String;
  stats6?: String;
  prop7?: String;
  min7?: String;
  max7?: String;
  stats7?: String;
  prop8?: String;
  min8?: String;
  max8?: String;
  stats8?: String;
  prop9?: String;
  min9?: String;
  max9?: String;
  stats9?: String;
  prop10?: String;
  min10?: String;
  max10?: String;
  stats10?: String;
  lineString: String;
  type?: String;
  tier?: String;
}

export interface IGrail {
  name: String;
  sets: ISetItem[];
  uniques: IUniqueItem[];
}

export interface IMule {
  name: String;
  userName: String;
  items: (ISetItem | IUniqueItem)[] | [];
}